python/WARA-PS_L2_Workshop_Agent/requirements.txt
paho-mqtt
haversine